+++ projects/pip-webui2-layouts/src/lib/nav-expander/nav-expander.module.ts
@@ -1,8 +1,7 @@
-import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatListModule } from '@angular/material/list';
+import { NgModule } from '@angular/core';
 import { MatIconModule } from '@angular/material/icon';
-
+import { MatListModule } from '@angular/material/list';
 import { PipNavExpanderComponent } from './nav-expander.component';
 
 @NgModule({